Add authentication (via /authenticate endpoint), add current_user to appctrl
ApplicationController can always tell the current user, the authentication
is not enforced, so auth errors are ignored (from JWT).
There is a new AuthenticatedController which enforces the authentication,
so if the user cannot be authenticated, the controller rejects with HTTP 401.
Creating the JWT a stateless session is created, so a SessionController
handles the request of /authenticate and for convinience the /login
endpoints.

Creating a user can be done eg:
User.create!(email: 'user@mail', password: 'secret', password_confirmation: 'secret')